--- a/addons/io_scene_gltf2/blender/imp/gltf2_blender_gltf.py
+++ b/addons/io_scene_gltf2/blender/imp/gltf2_blender_gltf.py
@@ -130,10 +130,6 @@ class BlenderGlTF():
         # Init is to False, and will be set to True during creation
         gltf.animation_object = False
 
-        # Store shapekeys equivalent between target & shapekey index
-        # For example when no POSITION on target
-        gltf.shapekeys = {}
-
         # Blender material
         if gltf.data.materials:
             for material in gltf.data.materials:
@@ -301,6 +297,33 @@ class BlenderGlTF():
             for mesh in gltf.data.meshes:
                 mesh.blender_name = None
                 mesh.is_weight_animated = False
+
+        # Calculate names for each mesh's shapekeys
+        for mesh in gltf.data.meshes:
+            mesh.shapekey_names = []
+            used_names = set()
+
+            for sk, target in enumerate(mesh.primitives[0].targets or []):
+                if 'POSITION' not in target:
+                    mesh.shapekey_names.append(None)
+                    continue
+
+                # Check if glTF file has some extras with targetNames. Otherwise
+                # use the name of the POSITION accessor on the first primitive.
+                shapekey_name = None
+                if mesh.extras is not None:
+                    if 'targetNames' in mesh.extras and sk < len(mesh.extras['targetNames']):
+                        shapekey_name = mesh.extras['targetNames'][sk]
+                if shapekey_name is None:
+                    if gltf.data.accessors[target['POSITION']].name is not None:
+                        shapekey_name = gltf.data.accessors[target['POSITION']].name
+                if shapekey_name is None:
+                    shapekey_name = "target_" + str(sk)
+
+                shapekey_name = BlenderGlTF.find_unused_name(used_names, shapekey_name)
+                used_names.add(shapekey_name)
+
+                mesh.shapekey_names.append(shapekey_name)
 
     @staticmethod
     def find_unused_name(haystack, desired_name):