@julienduroure 2019-09-25
Hello,
Thanks for the PR.
As you said, it will probably be reviewed only after 2.81 stabilization step.

"Unfortunately", I also have a big refactor of this part of the code, not really compatible with this one. We will have to get the best part of the both PR.
With my PR, some very big files go from 41 minutes to 1.5s :)