@julienduroure 2019-09-26
OK, I tested this PR today. Seems to be OK.
I will merge it now, because we are still in Bcon2.

I will then commit another importer perf improvement